Resolve S11P21C205-329 "Be/feature/accompany board  be 동행 api 테스트"

See merge request s11-bigdata-dist-sub1/S11P21C205!168

+++ backend/accompanyboard/src/main/java/com/ssafy/handam/accompanyboard/presentation/request/article/AccompanyBoardArticleCreationRequest.java
@@ -7,6 +7,6 @@ public record AccompanyBoardArticleCreationRequest(
         String description
 ) {
     public static AccompanyBoardArticleCreationRequest toServiceRequest(AccompanyBoardArticleCreationRequest request) {
-        return new AccompanyBoardArticleCreationRequest(request.userId(), request.scheduleId(), request.title(), request.description());
+        return new AccompanyBoardArticleCreationRequest(request.userId().longValue(), request.scheduleId().longValue(), request.title(), request.description());
     }
 }
